Update reusable backend related text to focus only on the Node JS project
