+++ .storybook/Welcome.mdx
@@ -21,7 +21,7 @@ This Storybook documentation provides a central place to view and interact with 
 
 - **Product List POC**: A product listing proof-of-concept that uses Supabase for backend support, showcasing modern React and Next.js techniques for managing cart items, applying discounts, and managing state.
 - **World of CSS Effects**: A collection of CSS techniques and animations that you can reuse for unique visual effects, including glitch effects, neon text, and more.
-- **Reusable Backend**: Links to reusable backend templates across different languages (Node JS, PHP, and C#) intended as modular, scalable backend solutions.
+- **Reusable Backend**: A reusable Node JS application template using modern backend solutions to create scalable and reusable modules with DDD
 - **Storybook Explorer**: This Storybook itself, configured with multiple addons and options to demonstrate its capabilities for creating organized, interactive UI documentation.
 
 ### Additional Repositories
